restructure: tier I — split vm.rs into vm/{opcode, exec}
